Cast the lm_head weight in the chunked backward, as the forward already does
_ChunkedLogProbFunction recomputes the logits chunk by chunk in backward, but
unlike forward it matmuls the lm_head weight without casting it to the hidden
dtype, into an `out` buffer allocated with the hidden dtype:

    forward:   w_chunk = weight[start:end].to(last_hidden.dtype)
               torch.mm(last_hidden, w_chunk.t(), out=mm_buf[:, :C])
    backward:  w_chunk = weight[start:end]
               torch.mm(hidden, w_chunk.t(), out=mm_buf[:, :C])

The comment above the forward cast names the case it exists for: under autocast
the hidden states are half precision while the lm_head weight is not. In that
case backward raises

    RuntimeError: expected m1 and m2 to have the same dtype,
                  but got: c10::BFloat16 != float

so a forward that succeeds cannot be backpropagated. The two same-dtype paths,
which is what the existing tests cover, are unaffected.

Cast only inside the matmul rather than rebinding w_chunk, so the gradient
accumulation below keeps the full precision weight; `.to()` on an already
matching dtype is a no-op, so the float32 and bfloat16 paths are unchanged.

test_backward_half_hidden_with_full_precision_weight covers bfloat16 and
float16 hidden states against a float32 weight, checking the returned gradient
dtypes and comparing both gradients against a reference that applies the same
cast.

+++ trl/trainer/utils.py
@@ -1427,7 +1427,11 @@ class _ChunkedLogProbFunction(torch.autograd.Function):
                 C = end - start
                 w_chunk = weight[start:end]  # [C, H]
 
-                torch.mm(hidden, w_chunk.t(), out=mm_buf[:, :C])
+                # Same cast as in forward: under autocast the hidden states are half precision while
+                # the lm_head weight is not, and `torch.mm` needs both operands and `out` in one dtype.
+                # Only the logit recompute is cast; the gradient accumulation below keeps the full
+                # precision weight.
+                torch.mm(hidden, w_chunk.to(hidden.dtype).t(), out=mm_buf[:, :C])
                 if bias is not None:
                     mm_buf[:, :C].add_(bias[start:end].to(hidden.dtype))
                 logits_chunk = logits_buf[:, :C]
